+++ runtime/xchainer.h
@@ -39,7 +39,7 @@ bool HasInf(const chainerx::Array& a);
 
 chainerx::Array Concat(const std::vector<chainerx::Array>& inputs, int axis);
 
-std::vector<chainerx::Array> Split(const chainerx::Array& input, const std::vector<int>& split, int axis);
+std::vector<chainerx::Array> Split(const chainerx::Array& input, const std::vector<int64_t>& split, int axis);
 
 chainerx::Array Pad(const std::vector<chainerx::Array>& inputs, int axis);
